requirements.txt
flask==2.0.1
joblib==1.0.1
pandas==1.3.0
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
werkzeug==2.0.1